requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
notion-client>=2.2.0
